pyproject.toml
[tool.poetry]
name = "advanced-codex-example"
version = "0.1.0"
description = "Advanced OpenAI API integration example in Python"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
asyncio = "^3.4.3"
aiohttp = "^3.8.1"
dependency-injector = "^4.41.0"